The design of an igniting agent for fusion energy systems
Single-layer graphene is currently produced on an ongoing basis to meet the demands of researchers. In recent years, there have been significant discoveries regarding the unique properties of this material, one of which is the generation of plasmons in graphene in response to excitement in the THz frequency range. Based on this concept, the present paper proposes two possible designs for a fusion energy ignition agent. These concepts could potentially allow fusion ignition at lower energies than are currently possible.

The basic concepts on which these designs are based are as follows.

1) The Coulomb forces associated with ionized deuterium and tritium nuclei are negated by the formation of higher energy plasmons around these nuclei.

2) The electric field has its highest value at the regions indicated by the yellow points in Figure 1. Consequently, ionized fuels would be expected to migrate towards the closest yellow points from both sides.