The design of an igniting agent for fusion energy systems ver2.5
Single-layer graphene is currently produced on an ongoing basis to meet the demands of researchers. In recent years, there have been significant discoveries regarding the unique properties of this material, one of which is the generation of plasmons in graphene in response to excitement in the THz frequency range. Based on this concept, the present paper proposes two possible designs for a fusion energy ignition agent. These concepts could potentially allow fusion ignition at lower energies than are currently possible.